feat(Syntax): add syntax highlighting
Adds syntax highlighting via highlight.js since that is what Quill uses.
Syntax highlighting can be enabled via the new syntax option by setting it to true.
syntax highlighting, when enabled, requires highlight.js be on the window as window.hljs (just like Quill) or it will throw.
Updated README to add syntax to options

BREAKING CHANGE: with this change the textarea becomes a Quill; always, reguardless of whether or not syntax is enabled.
Migration: If you have additional custom styles, they may need to be changed to target the quill classes instead of the textarea.

Closes #15