Allow to set where the overlayContainer is inserted
Currently I had a problem trying to open the html editor inside a
bootstrap modal. Bootstrap modal uses a display block property that
prevents the HTML editor allow to edit the content. I was able to solve
the problem inserting the overlayContainer before the bootstrap modal.

This adds a new option called prependSelector that allows to select
where you want to insert the overlayContainer.

Co-authored-by: Oscar Zatarain <ozatarain@hotmail.com>