Minor HTML tweaks
Currently the buttonOk doesn't have a CSS class for customization and
the popupTitle is an `i` tag that sets the font-style italic by
default. This change adds the ql-html-buttonOk for the buttonOk and
updates the popupTitle to a span class that is decorated with the
font-style italic to keep the current design.

Co-authored-by: Juan Zenon <juanjoze0424@gmail.com>